Fix tests around JSON filters
