+++ query-engine/connector-test-kit-rs/query-engine-tests/tests/queries/filters/mod.rs
@@ -10,6 +10,7 @@ pub mod filter_unwrap;
 pub mod filters;
 pub mod insensitive_filters;
 pub mod json;
+pub mod json_filter;
 pub mod json_path;
 pub mod list_filters;
 pub mod many_relation;
